Rewrote the entirety of the prolog.vim syntax file from posix to completely replace the default that comes with vim.
Instead of handcrafting regular expressions to define the prolog syntax,
the commit leverages the character sets directly from SWI-Prolog. The
builtins and keywords from various libraries are also included as
keywords for each match group. Support for directives that are typically
defined in a single-line were also defined along with any known options
that can be used with them.

Unfortunately, the options are presently disabled due to them requiring
usage of the "contains/contained" match keywords. In the future, if this
syntax plugin gets refactored to use a proper syntax hierarchy for
prolog's grammar, these keywords could be attached to the predicates
they are associated with.

Some significant improvements have been made to "prologOperators",
and "prologNumbers". This includes proper highlighting for rational
numbers, allowing usage of the "NaN" and "Inf" suffix with any real
numbers, and also including the "inf" and "nan" numbers as valid
numerical atoms. In the future, "inf" and "nan" should be bound to the
right-hand-side of the "is" keyword.

Operators have been completely refactored to separate the patterns used
for binary operators which are composed of multiple characters from
single-character operators that can be treated as binary or unary
depending on their context. Specifically, this is for distinguishing the
"+" and "-" operators so that they're only considered when they are
cuddled by atoms/variables/integers/compounds on both sides. Operators
composed of repeating characters have also been modified to match on
more than just a sequence. This way, operators will only be colored if
isolated from other symbols or distinctly separated by some whitespace.