Commented out the "EnableEscapeCommandline" option from the posix ssh client configuration.
For some stupid reason, the OpenSSH developers added this option to
newer versions but set it as disabled by default. Prior to the existence
of this option, the escape commandline had always existed. Thus, the
initial setting of disabled breaks backwards compatibility.

Since the OpenSSH client bails on an unknown option, it is impossible to
conditionally set the "EnableEscapeCommandline" option in a way that is
still compatible with both newer and older versions of OpenSSH. The only
way this can be remedied is by doing the version-checking outside of the
OpenSSH client configuration somehow. Until that happens, though, we
leave a description of the issue alongside the commented-out option.