Tweaked `lm` and `select_process_mappings` in the posix gdbinit file to fail with help.
This commit does a couple of things to improve usability and displays
the help if the first parameter for `select_process_mappings` is
incorrect. If address matching fails, the error message was tweaked to
emit the tested address with the number of digits matching the word
size.

The `lm` command had to be hacked up a bit so that you can pass too
little or too many parameters, which'll result in causing the
`select_process_mappings` command to fail with its help. I tried to
paste each of the arguments together as a string using `eval` and a loop
as suggested in the documentation to support variable-arguments, but
couldn't get it to work because user-defined functions really want to
evaluate their parameters. So instead I nested a bunch of conditionals
together because gdb doesn't have anything like else-if.