Fixed the tag matching use by `lxml.etree.iterparse` from the bin/html-select posix script.
Apparently `lxml.etree` can return tags that are prefixed with the xml
namespace that it assumes from the document type. This would result in
situations where no elements are returned despite the tags definitely
existing within the parsed file.

Instead, we use the tag the user gave us to produce a namespaced version
using a wildcard for matching, and a non-namespaced one. Although, I
definitely remember this already working before, so I'm not really sure
why this issue was never encountered. Anyways, now it is fixed, and
should be working hopefully forever.