Tweaked the "next" report in the taskwarrior configuration from the posix environment.
The "next" report when being displayed didn't always display the
"scheduled" column due to its "countdown" formatting. This commit
exchanges the "scheduled.countdown" formatting for "scheduled.relative".
This results in there always being a value displayed for that specific
column. It's apparent that I have no idea what the "countdown"
formatting for "scheduled" is actually supposed to do.