Added a new always-schedule hook to the taskwarrior configuration from the posix environment.
This new hook will ensure that the "SCHEDULED" field will always be set
to something when a task is created. This results in the column for the
task always having something displayed in it. When adding a new task,
the plugin checks if the "SCHEDULED" field hasn't been set. If so, the
plugin will extract both the "ENTRY" and "WAIT" fields and choose
whichever one is more recent when adding the "SCHEDULED" field.