Modified the barriers logged by the bin/for-idb.sh script from the posix environment.
The logfile emitted by the "bin/for-idb.sh` script includes so-called
"barriers" that can be used with gawk(1) to filter out unrelated lines
from the file. The barriers used are "^^^..." for starting execution of
the script with "$$$..." for ending or finishing execution of it.

However, the sister of this script, "bin/build-idb.sh", actually uses
three total barriers. The additional barrier it uses is "~~~..." which
is used to separate the output from the code related to it from the
code related to the python being executed.

To remedy this difference, this commit adds support for emitting the
"~~~..." barrier before saving or aborting. It also moves the finishing
barrier, "$$$...", to the point before the disassembler is terminated.