Tweaked the virtual-tag-blocks hook from the taskwarrior configuration.
This commit modifies the "virtual-tag-blocks" taskwarrior hook from the
posix environment so that it outputs any errors or exceptions that are
encountered as feedback that can be interpreted by the taskwarrior host
process. If the execution of the script was actually successful, it will
also emit feedback describing the tasks that were involved.