Modified the `annotation#property#get` implementation to always return a dictionary.
This commit fixes an issue with the "annotation" vim plugin that is
currently being developed within this repository. The autoload component
for getting the property for a location, `annotation#property#get`, is
supposed to return the dictionary of the first property that is found.
If no properties were found at said location, then the implementation
would mistakenly not return anything.

The issue was fixed by ensuring that a dictionary is always returned so
that it can be checked with `empty(...)` to determine if no properties
were found.