Added utility functions to the `annotation#state` autoloads for the vim annotation plugin.
The "annotation" vim plugin being developed within this repository
implements the `annotation#state` autoload with the intent that it can
be used to create and remove buffer state as well as track the
properties for a buffer and the data associated with them.

Due to an oversight, the implementation did not include any tools for
determining if the state for a buffer or its properties existed. This
commit fixes both by providing the `annotation#state#exists` function to
check buffer state, and the `annotation#state#hasprop` function to check
for the state associated with a property.