Updated `annotation#frontend` from the annotation plugin to consolidate the property data setting functions.
Originally, the `annotation#frontend#set_property_data` function was
written to take a parameter that gets applied to the data for the
specified property. To allow updating the property data based on the old
data, the `annotation#frontend#modify_property_data` function was added
which replaced the data with a "setter" using `funcref()`.

It was eventually realized that the login for both of these functions
can be combined into a single one that does explicit checks against the
type in order to determine which logic to use. Hence, this commit
replaces the `annotation#set_property_data` function with an
implementation that merges the functionality from both originals.