Tweaked the bashrc posix dotfile to deter login interactions on platforms configured for stupid people.
Some distros install the "PackageKit-command-not-found" package by
default which results in asking the user if they want to install a
package for a missing command. Since I'm using `command` to check if
a certain binary exists in the path, this results in asking the user on
login whether they want to install "devtodo". Despite me not using this
"devtodo" tool anymore, the missing-command shit is purely for stupid
people that don't know how to posix by default.

As a remedy, we add a random parameter to our usage of `command` to
prevent the "command_not_found_handle" bash function from executing. We
also remove the "command_not_found_handle" function during startup.