Updated the "Developer Console Extensions" browser userscript to add a tool for returning the stack trace.
Apparently by instantiating an `Error` object, its "stack" property
contains the backtrace from the scope of the instantiation. The only
caveat is that this backtrace is a straight-up string.

So, to simplify its access we split up the string by newlines, and then
split out both the line number and column number, the url, and the
function name (if available) from each line.  This way we can append
each discovered entry to a list that we can return. This functionality
was attached to the `console` object that is available during debugging.