Fixed an issue in the ssh client configuration for the posix environment.
The original ssh client configuration used the "Include" keyword to add
both site and host-specific options to the default client configuration.
However, due to the way that parsing for the `ssh_config(5)` actually
works, the files included by the "Include" keyword were being applied to
the loopback "Match" context that was being specified right before it.
This basically meant that all of the options from the site configuration
and host configuration directories were not being applied globally as
was originally intended.

Basically.. when `ssh(1)` parses its configuration file, it sets the
initial context to the host "*" so that when any options are being read,
they will be applied globally to all hosts. Then, when it encounters a
"Host" or "Match" keyword, it will the take next lines that follow to be
related to the host or whatever is being matched. So apparently, the
whitespace and indentations inside `ssh_config(5)` don't actually mean
anything.

To fix this issue, we change the context to the host "*" before we try
to include any site-specific or host-specific configurations. This
should result in anything being included to be applied as a global
option.