Modified the posix bin/fetch script to check the size before downloading a file.
Originally, the "download" function from the "bin/fetch" script would
start the download, and then monitor the file size for any data that has
been downloaded. This should be fine, but `curl(1)` will receive an HTTP
error if the resume offset is larger than the available file size.

This commit changes the order of operations so that the file size is
checked first, and then only after will the file be downloaded. Some
conditions were added for checking if the downloaded file size is larger
than the expected size, and for completing successfully when both of the
sizes actually match.