Tweaked the bin/fetch posix script to filter out errors when calculating checksums.
When an unsupported hash algorithm is used to calculate the signature,
the `openssl(1)` binary while write its complaints to the file
descriptor for `STDERR_FILENO`. This actually shouldn't happen because
we process the output of `openssl dgst -list` in order to determine
which algorithms are available. Despite this defense, `openssl(1)` can
actually lie about which algorithms are supported. This commit silences
the output written to the `STDERR_FILENO` for `openssl(1)` by
redirecting the file descriptor to "/dev/null".

Some of the output logs for the script were also rewritten to replace
instances of the word "download" with "fetch". This way the naming
better corresponds with the name of the actual script.