Fixed a bug in the "bin/cygpath" script from the posix environment.
In order to determine the current platform type, the "MACHTYPE"
variable as defined by `bash(1)` is used. Unfortunately on some versions
of `bash(1)`, the third component of the "MACHTYPE" uses "Cygwin"
instead of "Msys." This prevents the script from distinguishing whether
it is on the "Msys" or the "Cygwin" platforms.

This issue was remedied by using the "-o" parameter of `uname(1)`. So,
when choosing the output format we use `uname(1)` to get the operating
system, split up any unnecessary characters (like "/"), and then
lowercase the result. This way we can determine the correct method for
emitting the given path on the operating system currently being used.